VisualComponent/UI/python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
pydantic==2.5.3
ffmpeg-python==0.2.0
torch>=2.0.0
torchvision>=0.15.0
trimesh>=4.0.0
rembg>=2.0.50
omegaconf>=2.3.0
einops>=0.7.0
transformers>=4.35.0
diffusers>=0.25.0
pytest>=7.4.0
httpx>=0.25.0
